[#4798] Add option to conceal item descriptions from players
Adds a new setting, "Conceal Item Descriptions", which hides item
description in chat messages for players so long as they are (1)
posted by the GM and (2) don't have a chat description.

To avoid overloading the settings dialog any more, this adds a new
setting configuration app that now contains all of the player
visibility-releated settings, such as attack result visibility,
challenge visibility, and bloodied effect visibility. Some
reshuffling of the existing settings applications was done so they
are all now organized into the `applications/settings` folder.

A new `BaseSettingsConfig` application was introduced to make it
easy to add new settings applications that follow this pattern,
and the bastions config application was refactored to use this
new base application.

Closes #4798
