+++ lang/en.json
@@ -3767,20 +3767,6 @@
 "SETTINGS.5eAutoCollapseCardN": "Collapse Item Cards in Chat",
 "SETTINGS.5eAutoSpellTemplateL": "When a spell is cast, defaults to begin the process to create the corresponding Measured Template if any (requires TRUSTED or higher player role)",
 "SETTINGS.5eAutoSpellTemplateN": "Always place Spell Template",
-"SETTINGS.5eAttackRollVisibility": {
-  "Name": "Attack Result Visibility",
-  "Hint": "Control visibility of attack roll results in chat cards for players.",
-  "All": "Show results & target ACs",
-  "HideAC": "Show only results",
-  "None": "Hide all"
-},
-"SETTINGS.5eChallengeVisibility": {
-  "Name": "Challenge Visibility",
-  "Hint": "Control what roll DCs are visible to the players and whether successes/failures are highlighted.",
-  "All": "Show all",
-  "None": "Hide all",
-  "Player": "Show only from other players"
-},
 "SETTINGS.5eCriticalMaxDiceN": "Critical Damage Maximize Dice",
 "SETTINGS.5eCriticalMaxDiceL": "Make critical hits more deadly by maximizing the values of base damage dice.",
 "SETTINGS.5eCriticalModifiersN": "Critical Damage Multiply Modifiers",
@@ -3876,6 +3862,29 @@
     "Hint": "Change handling of various rules between the 2024 and 2014 rule sets.",
     "Legacy": "Legacy Rules (2014)",
     "Modern": "Modern Rules (2024)"
+  },
+  "VISIBILITY": {
+    "Hint": "Various configuration options that affect player visibility of certain information that the DM may wish to keep secret.",
+    "Label": "Configure Visibility",
+    "Name": "Visibility",
+    "ATTACK": {
+      "Name": "Attack Result Visibility",
+      "Hint": "Control visibility of attack roll results in chat cards for players.",
+      "All": "Show results & target ACs",
+      "HideAC": "Show only results",
+      "None": "Hide all"
+    },
+    "CHALLENGE": {
+      "All": "Show all",
+      "Hint": "Control what roll DCs are visible to the players and whether successes/failures are highlighted.",
+      "Name": "Challenge Visibility",
+      "None": "Hide all",
+      "Player": "Show only from other players"
+    },
+    "ITEMDESCRIPTIONS": {
+      "Hint": "Hide the descriptions of NPC items from players when posted to chat by the DM. Will only hide the descriptions for items without a Chat Description set.",
+      "Name": "Conceal NPC Descriptions"
+    }
   }
 },
