+++ module/applications/_module.mjs
@@ -9,13 +9,12 @@ export * as fields from "./fields.mjs";
 export * as item from "./item/_module.mjs";
 export * as journal from "./journal/_module.mjs";
 export * as mixins from "./mixins/_module.mjs";
+export * as settings from "./settings/_module.mjs";
 export * as shared from "./shared/_module.mjs";
 
 export {default as Accordion} from "./accordion.mjs";
 export {default as Award} from "./award.mjs";
-export {default as BastionConfig} from "./bastion.mjs";
 export {default as CompendiumBrowser} from "./compendium-browser.mjs";
-export {default as CompendiumBrowserSourceConfig} from "./compendium-browser-source-config.mjs";
 export {default as ContextMenu5e} from "./context-menu.mjs";
 export {default as CurrencyManager} from "./currency-manager.mjs";
 export {default as DialogMixin} from "./dialog-mixin.mjs";
@@ -24,3 +23,7 @@ export {default as SheetConfig} from "./sheet-config.mjs";
 export {default as SourceConfig} from "./source-config.mjs";
 export {default as Tabs5e} from "./tabs.mjs";
 export {default as TokenConfig5e} from "./token-config.mjs";
+
+// Deprecations
+export {default as BastionConfig} from "./bastion.mjs";
+export {default as CompendiumBrowserSourceConfig} from "./compendium-browser-source-config.mjs";
