+++ module/applications/settings/_module.mjs
@@ -1,0 +1,5 @@
+export {default as BaseSettingsConfig} from "./base-settings.mjs";
+export {default as BastionSettingsConfig} from "./bastion-settings.mjs";
+export {default as CompendiumBrowserSettingsConfig} from "./compendium-browser-settings.mjs";
+export {default as ModuleArtSettingsConfig} from "./module-art-settings.mjs";
+export {default as VisibilitySettingsConfig} from "./visibility-settings.mjs";
