+++ module/settings.mjs
@@ -1,7 +1,7 @@
-import { ModuleArtConfig } from "./module-art.mjs";
-import CompendiumBrowserSourceConfig from "./applications/compendium-browser-source-config.mjs";
-import BastionConfig, { BastionSetting } from "./applications/bastion.mjs";
-import * as Bastion from "./documents/actor/bastion.mjs";
+import BastionSettingsConfig, { BastionSetting } from "./applications/settings/bastion-settings.mjs";
+import CompendiumBrowserSettingsConfig from "./applications/settings/compendium-browser-settings.mjs";
+import ModuleArtSettingsConfig from "./applications/settings/module-art-settings.mjs";
+import VisibilitySettingsConfig from "./applications/settings/visibility-settings.mjs";
 
 /**
  * Register all of the system's keybindings.
@@ -38,32 +38,31 @@ export function registerSystemSettings() {
     default: ""
   });
 
-  // Challenge visibility
   game.settings.register("dnd5e", "challengeVisibility", {
-    name: "SETTINGS.5eChallengeVisibility.Name",
-    hint: "SETTINGS.5eChallengeVisibility.Hint",
+    name: "SETTINGS.DND5E.VISIBILITY.CHALLENGE.Name",
+    hint: "SETTINGS.DND5E.VISIBILITY.CHALLENGE.Hint",
     scope: "world",
-    config: true,
+    config: false,
     default: "player",
     type: String,
     choices: {
-      all: "SETTINGS.5eChallengeVisibility.All",
-      player: "SETTINGS.5eChallengeVisibility.Player",
-      none: "SETTINGS.5eChallengeVisibility.None"
+      all: "SETTINGS.DND5E.VISIBILITY.CHALLENGE.All",
+      player: "SETTINGS.DND5E.VISIBILITY.CHALLENGE.Player",
+      none: "SETTINGS.DND5E.VISIBILITY.CHALLENGE.None"
     }
   });
 
   game.settings.register("dnd5e", "attackRollVisibility", {
-    name: "SETTINGS.5eAttackRollVisibility.Name",
-    hint: "SETTINGS.5eAttackRollVisibility.Hint",
+    name: "SETTINGS.DND5E.VISIBILITY.ATTACK.Name",
+    hint: "SETTINGS.DND5E.VISIBILITY.ATTACK.Hint",
     scope: "world",
-    config: true,
+    config: false,
     default: "none",
     type: String,
     choices: {
-      all: "SETTINGS.5eAttackRollVisibility.All",
-      hideAC: "SETTINGS.5eAttackRollVisibility.HideAC",
-      none: "SETTINGS.5eAttackRollVisibility.None"
+      all: "SETTINGS.DND5E.VISIBILITY.ATTACK.All",
+      hideAC: "SETTINGS.DND5E.VISIBILITY.ATTACK.HideAC",
+      none: "SETTINGS.DND5E.VISIBILITY.ATTACK.None"
     }
   });
 
@@ -71,7 +70,7 @@ export function registerSystemSettings() {
     name: "SETTINGS.DND5E.BLOODIED.Name",
     hint: "SETTINGS.DND5E.BLOODIED.Hint",
     scope: "world",
-    config: true,
+    config: false,
     default: "player",
     type: String,
     choices: {
@@ -390,7 +389,7 @@ export function registerSystemSettings() {
     label: "DND5E.ModuleArtConfigL",
     hint: "DND5E.ModuleArtConfigH",
     icon: "fa-solid fa-palette",
-    type: ModuleArtConfig,
+    type: ModuleArtSettingsConfig,
     restricted: true
   });
 
@@ -413,7 +412,7 @@ export function registerSystemSettings() {
     label: "DND5E.CompendiumBrowser.Sources.Label",
     hint: "DND5E.CompendiumBrowser.Sources.Hint",
     icon: "fas fa-book-open-reader",
-    type: CompendiumBrowserSourceConfig,
+    type: CompendiumBrowserSettingsConfig,
     restricted: true
   });
 
@@ -431,7 +430,7 @@ export function registerSystemSettings() {
     label: "DND5E.Bastion.Configuration.Label",
     hint: "DND5E.Bastion.Configuration.Hint",
     icon: "fas fa-chess-rook",
-    type: BastionConfig,
+    type: BastionSettingsConfig,
     restricted: true
   });
 
@@ -446,6 +445,25 @@ export function registerSystemSettings() {
       duration: 7
     },
     onChange: () => game.dnd5e.bastion.initializeUI()
+  });
+
+  // Visibility Settings
+  game.settings.registerMenu("dnd5e", "visibilityConfiguration", {
+    name: "SETTINGS.DND5E.VISIBILITY.Name",
+    label: "SETTINGS.DND5E.VISIBILITY.Label",
+    hint: "SETTINGS.DND5E.VISIBILITY.Hint",
+    icon: "fas fa-eye",
+    type: VisibilitySettingsConfig,
+    restricted: true
+  });
+
+  game.settings.register("dnd5e", "concealItemDescriptions", {
+    name: "SETTINGS.DND5E.VISIBILITY.ITEMDESCRIPTIONS.Name",
+    hint: "SETTINGS.DND5E.VISIBILITY.ITEMDESCRIPTIONS.Hint",
+    scope: "world",
+    config: false,
+    default: false,
+    type: Boolean
   });
 
   // Primary Group
