+++ module/utils.mjs
@@ -547,6 +547,7 @@ export async function preloadHandlebarsTemplates() {
     "systems/dnd5e/templates/shared/fields/field-range.hbs",
     "systems/dnd5e/templates/shared/fields/field-targets.hbs",
     "systems/dnd5e/templates/shared/fields/field-uses.hbs",
+    "systems/dnd5e/templates/shared/fields/fieldlist.hbs",
 
     // Journal Partials
     "systems/dnd5e/templates/journal/parts/journal-legacy-traits.hbs",
