--- templates/apps/bastion-config.hbs
@@ -1,12 +1,0 @@
-<fieldset>
-    <legend>{{ localize "SETUP.Configuration" }}</legend>
-
-    {{!-- Enabled --}}
-    {{ formField fields.enabled value=source.enabled input=inputs.createCheckboxInput localize=true }}
-
-    {{!-- Button --}}
-    {{ formField fields.button value=source.button input=inputs.createCheckboxInput localize=true }}
-
-    {{!-- Duration --}}
-    {{ formField fields.duration value=source.duration localize=true }}
-</fieldset>
