+++ templates/chat/activity-card.hbs
@@ -1,5 +1,5 @@
 <div class="dnd5e2 chat-card activation-card">
-    <section class="card-header description collapsible">
+    <section class="card-header description collapsible" {{#if description.concealed}}data-concealed{{/if}}>
         <header class="summary">
             <img class="gold-icon" src="{{ item.img }}" alt="{{ item.name }}">
             <div class="name-stacked border">
@@ -11,7 +11,7 @@
             <i class="fa-solid fa-chevron-down fa-fw" inert></i>
         </header>
         <section class="details collapsible-content card-content">
-            <div class="wrapper">{{{ description }}}</div>
+            <div class="wrapper">{{{ description.chat }}}</div>
         </section>
     </section>
