+++ templates/settings/base-config.hbs
@@ -1,0 +1,4 @@
+<fieldset>
+    <legend>{{ localize "SETUP.Configuration" }}</legend>
+    {{> "dnd5e.fieldlist" fields}}
+</fieldset>
