+++ templates/shared/fields/fieldlist.hbs
@@ -1,0 +1,3 @@
+{{#each this}}
+{{ formField field name=name value=value input=input options=options rootId=@root.partId }}
+{{/each}}
